requirements.txt
Django
djangorestframework
pytest
pytest-django
psycopg2-binary
requests
requests_mock
python-dotenv
dj-database-url
isort
ruff
black